requirements.txt

ipython
jupyter
invoke
scandir
ujson
numpy
scipy
matplotlib
pandas
SQLAlchemy
click
attrs
wordfreq
gensim
pandas
